# Fix: rewrite overwriting logic of config