--- a/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/.eslintrc
+++ b/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/.eslintrc
@@ -1,5 +1,8 @@
 {
     "ecmaFeatures": {
         "globalReturn": false
+    },
+    "globals": {
+        "require": true
     }
 }