--- a/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child/child/.eslintrc
+++ b/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child/child/.eslintrc
@@ -1,0 +1,5 @@
+{
+    "env": {
+        "commonjs": false
+    }
+}