--- a/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child2/.eslintrc
+++ b/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child2/.eslintrc
@@ -1,0 +1,6 @@
+{
+    "env": {
+        "node": true,
+        "commonjs": true
+    }
+}