--- a/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child2/both-false/.eslintrc
+++ b/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child2/both-false/.eslintrc
@@ -1,0 +1,6 @@
+{
+    "env": {
+        "node": false,
+        "commonjs": false
+    }
+}