--- a/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child3/.eslintrc
+++ b/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child3/.eslintrc
@@ -1,0 +1,6 @@
+{
+    "env": {
+        "commonjs": true,
+        "node": true
+    }
+}