--- a/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child3/child/.eslintrc
+++ b/tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child3/child/.eslintrc
@@ -1,0 +1,6 @@
+{
+    "extends": "test1",
+    "globals": {
+        "__dirname": true
+    }
+}