… lib/config.js
… tests/fixtures/config-hierarchy/overwrite-ecmaFeatures/child2/node-false/.eslintrc
… tests/lib/config.js