Fixes #3915 

I rewrite config's inheritance and overwriting logic.
The old way could not handle overwriting of environments with `false`.
The new way stores history of overwriting for each `globals` and `ecmaFeatures`. Then it selects the correct value from final `env` and history.
I added a class `CascadingConfigData` to manages history of overwriting.

----

This is a little bit faster than before.

**Before:**

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  3486.3081ms
Performance Run #2:  3462.279093ms
Performance Run #3:  3428.105712ms
Performance Run #4:  3418.651979ms
Performance Run #5:  3418.054008ms
Performance budget exceeded: 3428.105712ms (limit: 2211.0849056603774ms)
```

**After:**

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  3495.445491ms
Performance Run #2:  3406.694375ms
Performance Run #3:  3398.599885ms
Performance Run #4:  3395.976783ms
Performance Run #5:  3395.71085ms
Performance budget exceeded: 3398.599885ms (limit: 2211.0849056603774ms)
```