@mysticatea 2015-09-29
I had forgotten about `/* eslint-env` comments...