@nzakas 2015-09-29
Yikes! This is a pretty massive change. In the future, before you rewrite something this large, please talk about what you're planning on doing.

It's going to take me a while to review this.