@mysticatea 2015-09-30
I'm sorry. An intention of [this comment](https://github.com/eslint/eslint/issues/3915#issuecomment-142766786) was to make a large change. I will write more details in the future.

The core idea of this PR is [getSnapshot function](https://github.com/eslint/eslint/pull/3958/files#diff-af01dc6354b01e23c32eba0651e29428R228)
`CascadingConfigData` is a class for that.
`CascadingConfigData` instance exists for each config file (and shared config), the instances have tree structure.
And other changes are processes to make the tree structure of `CascadingConfigData`.

I will add a support of `/* eslint-env` comments to this PR later.
I will add a overload of `eslint.verify()` to accept `CascadingConfigData` instance in the argument `config`. `eslint.verify()` does exactly overwrite env with using `CascadingConfigData` if `config` is a `CascadingConfigData` instance. Otherwise, it does overwrite env as the same way as before.
Then I will make `CLIEngine` giving a `CascadingConfigData` instance to `eslint.verify()`.