@nzakas 2015-09-30
I think that's too much change to fix a bug. I'd like to see if there's a much smaller change to address this.