@mysticatea 2015-10-16
I don't have other idea yet.
I believe history of overwriting is needed to overwrite config correctly. My thinking cannot leave from there....

Now, yes, management of history is a large responsibility,  so I thought the responsibility should be extracted to a specialized class: `CascadingConfigData`.
This object manages history of overwriting as singly linked lists. And this resolves config at finally.
This class is isolated from file system.

`config.js` finds config files and makes `CascadingConfigData` from positional relationship of config files.

I considered the following patterns:

----

**src/.eslintrc**

```js
{
    "ecmaFeature": {"globalReturn": false}
}
```

**src/foo/.eslintrc**

```js
{
    "env": {"node": true}
}
```

* The result of `globalReturn` should be `true`.
  The way that resolves `env` content at finally cannot make this because existing `ecmaFeatures` is not overwritten. #3836. Now eslint resolves `env` content at each config and overwrites upper's. It makes correct in this case.

----

**src/.eslintrc**

```js
{
    "env": {"node": true}
}
```

**src/foo/.eslintrc**

```js
{
    "ecmaFeature": {"globalReturn": false}
}
```

* The result of `globalReturn` should be `false`.
  This is an opposite side of the previous pattern. The current way makes correct.

----

**src/.eslintrc**

```js
{
    "env": {"node": true, "commonjs": true}
}
```

**src/foo/.eslintrc**

```js
{
    "env": {"node": false}
}
```

* The result of `require` variable should exist.
  Such as `Buffer` should be removed, but `require` should not be removed. So we cannot remove `env` content merely. The current way cannot address this pattern correctly. But the previous way (to resolve `env` content at finally) can make correct.

----

**src/.eslintrc**

```js
{
    "ecmaFeature": {"globalReturn": true},
    "env": {"node": true}
}
```

**src/foo/.eslintrc**

```js
{
    "env": {"node": false}
}
```

* The result of `globalReturn` should be `true`.
  This is similar to the previous pattern.