@hsablonniere 2015-10-20
I didn't get into the details of the PR but we would really like to do this in our project ;-)