@nzakas 2015-12-08
I think we need to rethink this a bit. Closing for now.